refactor: rename repositories REST/GitHub
