+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/rest/repository/adapters/RepositoryGitHubRepository.java
@@ -1,36 +1,34 @@
 package io.pakland.mdas.githubstats.infrastructure.rest.repository.adapters;
 
-import io.pakland.mdas.githubstats.application.dto.UserDTO;
+import io.pakland.mdas.githubstats.application.dto.RepositoryDTO;
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
 import io.pakland.mdas.githubstats.infrastructure.rest.repository.WebClientConfiguration;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.ports.IUserRESTRepository;
+import io.pakland.mdas.githubstats.infrastructure.rest.repository.ports.IRepositoryRESTRepository;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Repository;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
 
-import java.util.List;
-
-@Repository
-public class UserRESTRepository implements IUserRESTRepository {
+public class RepositoryGitHubRepository implements IRepositoryRESTRepository {
 
     private final WebClientConfiguration webClientConfiguration;
-    Logger logger = LoggerFactory.getLogger(UserRESTRepository.class);
+    Logger logger = LoggerFactory.getLogger(OrganizationGitHubRepository.class);
 
-    public UserRESTRepository(WebClientConfiguration webClientConfiguration) {
+    public RepositoryGitHubRepository(WebClientConfiguration webClientConfiguration) {
         this.webClientConfiguration = webClientConfiguration;
     }
 
     @Override
-    public List<UserDTO> fetchUsersFromTeam(String organizationName, String teamSlug) throws HttpException {
+    public List<RepositoryDTO> fetchTeamRepositories(Integer orgId, Integer teamId)
+        throws HttpException {
         try {
             return this.webClientConfiguration.getWebClient().get()
-                    .uri(String.format("/orgs/%s/teams/%s/members", organizationName, teamSlug))
+                .uri(String.format("/organizations/%s/team/%d/repos", orgId, teamId))
                     .retrieve()
-                    .bodyToFlux(UserDTO.class)
+                    .bodyToFlux(RepositoryDTO.class)
                     .collectList()
                     .block();
-        }  catch (WebClientResponseException ex) {
+        } catch (WebClientResponseException ex) {
             logger.error(ex.toString());
             throw new HttpException(ex.getRawStatusCode(), ex.getMessage());
         }
